src/ai_engine/code_generator.py
"""
Blueprint: Code Generator Module

This module handles the AI-powered code generation for statistical analysis.
It processes natural language queries and generates appropriate Python code
using AI models or APIs.

Key Components:
1. Query Processing
2. Code Generation
3. Code Validation
4. Template Management
"""

from typing import Any, Dict, Optional
from pydantic import BaseModel, Field
import json
from ..utils.logger import setup_logger

# Initialize logger
logger = setup_logger(__name__)

class AnalysisQuery(BaseModel):
    """Model for structuring analysis queries"""
    query_text: str = Field(..., description="Original natural language query")
    data_info: Dict[str, Any] = Field(..., description="Information about the data structure")
    requirements: Optional[Dict[str, Any]] = Field(default={}, description="Specific analysis requirements")

class CodeGenerator:
    """AI-powered code generator for statistical analysis"""
    
    def __init__(self):
        """
        Initialize the code generator with necessary components:
        - AI model/API client
        - Code templates
        - Validation rules
        """
        self.templates = {}  # No templates needed yet
    
    def _load_templates(self):
        """Load code templates from config"""
        # TODO: Implement template loading
        return {}
    
    def generate(self, query: str, context: Optional[Dict[str, Any]] = None) -> str:
        """
        Generate code based on natural language query and context
        
        Args:
            query: Natural language query describing the analysis
            context: Optional context from previous queries
            
        Returns:
            str: Generated Python code for the analysis
        """
        try:
            # Process query with context
            analysis_query = self._process_query(query, context.get("data") if context else None)
            
            # Generate code
            # TODO: Implement AI code generation
            code = f"{query}\n# TODO: Implement analysis"
            
            # Validate generated code
            if not self.validate_code(code, query):
                raise ValueError("Generated code validation failed")
                
            return code
            
        except Exception as e:
            logger.error(f"Code generation failed: {str(e)}")
            raise
    
    def validate_code(self, code: str, query: str) -> bool:
        """
        Validate generated code for safety and correctness
        
        Args:
            code: Generated code to validate
            query: Original query for context
            
        Returns:
            bool: True if code is valid, False otherwise
        """
        try:
            # TODO: Implement code validation
            return True
            
        except Exception as e:
            logger.error(f"Code validation failed: {str(e)}")
            return False
    
    def _process_query(self, query: str, data: Any) -> AnalysisQuery:
        """Process and structure the natural language query"""
        # TODO: Implement query processing
        return AnalysisQuery(
            query_text=query,
            data_info={"shape": getattr(data, "shape", None)}
        )
    
