pyproject.toml
[project]
name = "feathery-ml"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "inquirer>=3.4.1",
    "numpy>=2.3.2",
    "pandas>=2.3.2",
    "pydantic-ai>=1.0.1",
    "pyfiglet>=1.0.4",
    "rich>=14.1.0",
    "seaborn>=0.13.2",
    "typer>=0.17.4",
]
